feat: refactor devimint into a library
this can be external integrators to run tests with custom modules
